% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_func.R
\name{first_func}
\alias{first_func}
\title{My First Function}
\usage{
first_func(x)
}
\arguments{
\item{x}{This is unused.}
}
\description{
This function does almost nothing.
}
\examples{
first_func('test')
}
